Move mixin methods to proper locations
This change moves methods for more efficient loading
and requires less reinstantiation.
