+++ directord/tests/test_init.py
@@ -233,24 +233,6 @@ class TestProcessor(unittest.TestCase):
                     self.log.debug.called_once()
         self.assertListEqual(chunks, ["d", "a", "t", "a"])
 
-    def test_file_sha1(self):
-        with unittest.mock.patch(
-            "builtins.open", unittest.mock.mock_open(read_data=b"data")
-        ) as mock_file:
-            sha1 = self.processor.file_sha1(file_path=mock_file)
-            self.assertEqual(sha1, "a17c9aaa61e80a1bf71d0d850af4e5baa9800bbd")
-
-    def test_file_sha1_set_chunk(self):
-        with unittest.mock.patch(
-            "builtins.open", unittest.mock.mock_open(read_data=b"data")
-        ) as mock_file:
-            sha1 = self.processor.file_sha1(file_path=mock_file, chunk_size=1)
-            self.assertEqual(sha1, "a17c9aaa61e80a1bf71d0d850af4e5baa9800bbd")
-
-    def test_object_sha1(self):
-        sha1 = self.processor.object_sha1(obj={"test": "value"})
-        self.assertEqual(sha1, "4e0b1f3b9b1e08306ab4e388a65847c73a902097")
-
     def test_run_threads(self):
         pass
